# course-work/Implementations/DSCourseProject/VideoLibraryAPI/wwwroot/Swagger/OpenAPISpec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VideoLibrary API",
    "description": "An ASP.NET Core Web API for managing videos in a BLOB storage \n\n Use https://editor-next.swagger.io/ when checking the documentation since its writen in version 3.1.0",

    "version": "v1"
  },
...
  "paths": {
...
    "/api/Video": {
      "post": {
        "tags": [
          "Video"
        ],
        "security": [
          {
            "EntraID": []
          }
        ],
        "summary": "Creates a new Video",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoIM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Vid Test Add"
                    },
                    "fileLocation": {
                      "type": "string",
                      "example": "Vid Test Add"
                    },
                    "fileID": {
                      "type": "string",
                      "format": "uuid",
                      "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "formatId": {
                      "type": "integer",
                      "example": 8
                    },
                    "format": {
                      "type": [ "string", "null" ],
                      "nullable": true,
                      "example": null
                    },
                    "size": {
                      "type": "integer",
                      "example": 10
                    },
                    "description": {
                      "type": "string",
                      "example": "Vid Test Add"
                    },
                    "yearOfPublishing": {
                      "type": "string",
                      "format": "date-time",
                      "example": "1980-01-01T00:00:00Z"
                    },
                    "copyrightId": {
                      "type": "integer",
                      "example": 6
                    },
                    "copyright": {
                      "type": [ "string", "null" ],
                      "nullable": true,
                      "example": null
                    },
                    "id": {
                      "type": "integer",
                      "example": 6
                    }
                  },
                  "required": [
                    "title",
                    "fileLocation",
                    "fileID",
                    "formatId",
                    "size",
                    "description",
                    "yearOfPublishing",
                    "copyrightId",
                    "id"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "One or more validation errors occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "One or more validation errors occurred."
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "errors": {
                      "type": "object",
                      "properties": {
                        "Title": {
                          "type": "array",
                          "items": { "type": "string" },
                          "example": [ "This field is Required" ]
                        },
                        "CopyrightId": {
                          "type": "array",
                          "items": { "type": "string" },
                          "example": [ "The field CopyrightId must be between 1 and 6." ]
                        },
                        "Description": {
                          "type": "array",
                          "items": { "type": "string" },
                          "example": [ "This field is Required" ]
                        },
                        "FileLocation": {
                          "type": "array",
                          "items": { "type": "string" },
                          "example": [ "This field is Required" ]
                        }
                      }
                    },
                    "traceId": {
                      "type": "string",
                      "example": "00-825bb4134c7fae6a27e3921d6e792542-3630ef2fadf1076b-00"
                    }
                  },
                  "required": [ "title", "status", "errors", "traceId" ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Video/get/author/{videoId}": {
      "post": {
        "tags": [
          "Video-Authors"
        ],
        "security": [
          {
            "EntraID": []
          }
        ],
        "summary": "Gets a list of Authors with connection with the specified video",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "ID of the video",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorVideoAuthorFilterIMIndexIM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Video/get/genre/{videoId}": {
      "post": {
        "tags": [
          "Video-Genres"
        ],
        "security": [
          {
            "EntraID": []
          }
        ],
        "summary": "Gets a list of Genres with connection with the specified Video",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagVideoGenreFilterIMIndexIM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Video/get/tag/{videoId}": {
      "post": {
        "tags": [
          "Video-Tags"
        ],
        "security": [
          {
            "EntraID": []
          }
        ],
        "summary": "Gets a list of Tags with connection with the specified Video",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagVideoTagFilterIMIndexIM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorVideoAuthorFilterIMIndexIM": {
        "type": "object",
        "description": "Used as input: sends Paging and Filtering to API \n\n Used as output: returns retrieved items, Filter and paging information",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorVideo"
            },
            "nullable": true
          },
          "pager": {
            "$ref": "#/components/schemas/PagerIM"
          },
          "filter": {
            "$ref": "#/components/schemas/AuthorFilterIM"
          }
        },
        "additionalProperties": false
      },
...
      "PagerIM": {
        "type": "object",
        "description": "Contains information used for paging",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "examples": [ 1, 2, 100 ],
            "description": "The current page number",
            "minimum": 1,
            "maximum": 2147483647
          },
          "itemsPerPage": {
            "type": "integer",
            "format": "int32",
            "examples": [ 10, 20, 50 ],
            "description": "How many items are represented per 1 page",
            "minimum": 1,
            "maximum": 2147483647
          },
          "pagesCount": {
            "type": "integer",
            "format": "int32",
            "examples": [ 12, 20, 100 ],
            "description": "How many pages there are",
            "minimum": 1,
            "maximum": 2147483647
          }
        },
        "additionalProperties": false
      },
...
      "TagVideoGenreFilterIMIndexIM": {
        "type": "object",
        "description": "Used as input: sends Paging and Filtering to API \n\n Used as output: returns retrieved items, Filter and paging information",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagVideo"
            },
            "nullable": true
          },
          "pager": {
            "$ref": "#/components/schemas/PagerIM"
          },
          "filter": {
            "$ref": "#/components/schemas/GenreFilterIM"
          }
        },
        "additionalProperties": false
      },
...
      "TagVideoTagFilterIMIndexIM": {
        "type": "object",
        "description": "Used as input: sends Paging and Filtering to API \n\n Used as output: returns retrieved items, Filter and paging information",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagVideo"
            },
            "nullable": true
          },
          "pager": {
            "$ref": "#/components/schemas/PagerIM"
          },
          "filter": {
            "$ref": "#/components/schemas/TagFilterIM"
          }
        },
        "additionalProperties": false
      },
...
      "VideoIM": {
        "required": [
          "copyrightId",
          "description",
          "fileId",
          "formatId",
          "size",
          "title",
          "yearOfPublishing"
        ],
        "type": "object",
        "description": "Input data for actions that alter the database",
        "properties": {
          "title": {
            "minLength": 1,
            "maxLength": 50,
            "type": "string"
          },
          "fileId": {
            "minLength": 1,
            "type": "string"
          },
          "size": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "formatId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "minLength": 1,
            "maxLength": 255,
            "type": "string"
          },
          "yearOfPublishing": {
            "type": "string",
            "format": "date-time"
          },
          "copyrightId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "EntraID": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com/f68e025c-940a-4eaf-8cc3-d486d95844a2/oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com/f68e025c-940a-4eaf-8cc3-d486d95844a2/oauth2/v2.0/token",
            "scopes": {
              "api://e6de894a-7857-4ac4-babe-bb5731887f9f/Weather.Get": "Access the API as the signed-in user"
            }
          }
        }
      }
    },
...
  }
}